+++ src/countries/iq.ts
@@ -5,6 +5,21 @@ export const iq: Country = {
 	name: {
 		en: 'Iraq',
 		ar: 'العراق',
+		el: 'Ιράκ',
+		sv: 'Irak',
+		tr: 'Irak',
+		th: 'อิรัก',
 	},
 	flag: '🇮🇶',
+	nativeLanguage: 'ar',
+	anthem: {
+		nativeName: 'موطني',
+		name: {
+			en: 'My Homeland',
+			ar: 'موطني',
+		},
+		intro: 4.3,
+		composed: '1934',
+		adopted: '2004',
+	},
 }